requirements.txt
streamlit==1.12.0
numpy==1.19.5
Pillow==8.3.2
scikit-learn==0.23.0
tensorflow-cpu==2.4.1
